// src/generated/model/carType.ts
/**
 * OpenAPI ZbW car rent
 * zbw car rent api
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * A car type class
 */
export interface CarType { 
    idCarType?: number;
    carType?: string;
}